Cannot get the indent to work
The `withBlock` function should parse all lines in the following block
that are indented beyond the reference. However, I am finding that only
the first level of indentation is being handled.

So this works:

```
source R:
    foo()
    bar()
```

But this fails:

```
source R:
    foo(x){
        x
    }
```

It fails because it expects `x` to be a statement.