Add in Parsec.Language lexer and semicolons
Now I use `Text.Parsec.Language` to build a lexer for parsing strings,
comments, variables, etc. Now that I have a better idea of what it is
doing, it is pretty nice to have.

Also, all statements now must end in a semicolon. This is a little ugly,
but it is easier to parse than the Haskell/Python style indentation
syntax. I will consider this a temporary change.

I also defined data structures for expressions, but have not yet
implemented the parsers.